Remove duration as a public property type
The mapping to std::int64_t and i64 isn't really idiomatic, it would be better
to use std::chrono::time and std::time::duration. But those can't be used directly but would have to
available via conversion from an opaque sixtyfps::Duration type, to
maintain ffi compatibility (allow creation in C++, pass to Rust runtime).

For now we decided to remove the inconvenient mapping from the public API and
introduce the better one later when we need it.